libs/KPReID/requirements.txt
# base torchreid
numpy
Cython
h5py
Pillow
six
scipy
opencv-python
matplotlib
tb-nightly
future
yacs
gdown
flake8
yapf
isort

# added for bpbreid
albumentations==1.3.1
pandas
tabulate
deepdiff
wandb
monai
torchmetrics==1.3.0

# added for kpr
git+https://github.com/facebookresearch/segment-anything.git
git+https://github.com/katsura-jp/pytorch-cosine-annealing-with-warmup
timm
scikit-image
tqdm
yacs
omegaconf
scikit-learn
